requirements.txt
robotframework==6.0.2
Appium-Python-Client==-2.8.1
selenium==4.8.0
requests==2.28.2
allure-robotframework==2.12.0
chromedriver-autoinstaller==0.4.0
undetected-chromedriver==3.4.4
edgedriver-autoinstaller==0.2.1
undetected-edgedriver==3.1.6.4
